py/requirements.txt
essentia>=2.1b.dev532
music21>=5.5.0
sound-machine>=1.0
music>=0.1b0